QPID-8144: [Broker-J] Correct arithmetic type used when computing utilised direct memory (from int to long).
Cherry picked from master 8e6d2c1f2cb47df31d7d2ccb8e00dfed6b713a70

+++ broker-core/src/main/java/org/apache/qpid/server/bytebuffer/QpidByteBufferFactory.java
@@ -366,7 +366,7 @@ final class QpidByteBufferFactory
 
     static long getAllocatedDirectMemorySize()
     {
-        return _pooledBufferSize * getNumberOfBuffersInUse();
+        return (long)_pooledBufferSize * getNumberOfBuffersInUse();
     }
 
     static int getNumberOfBuffersInUse()
